# internal/adapter/postgres/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type OauthIdentity struct {
	ID              string             `json:"id"`
	UserID          string             `json:"user_id"`
	Provider        string             `json:"provider"`
	ProviderSubject string             `json:"provider_subject"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type RefreshToken struct {
	ID        string             `json:"id"`
	UserID    string             `json:"user_id"`
	TokenHash string             `json:"token_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	RevokedAt pgtype.Timestamptz `json:"revoked_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type User struct {
	ID            string             `json:"id"`
	Email         string             `json:"email"`
	DisplayName   string             `json:"display_name"`
	AvatarUrl     string             `json:"avatar_url"`
	EmailVerified bool               `json:"email_verified"`
	PasswordHash  *string            `json:"password_hash"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
}